Make Quaternion and Vector3 inner classes in MPU6050
